# swagger.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/v1/products/":{"get":{"summary":"Get Products","operationId":"get_products_v1_products__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductBase"},"title":"Response Get Products V1 Products  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Product","operationId":"create_product_v1_products__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/products/{product_id}/price":{"patch":{"summary":"Update Price","operationId":"update_price_v1_products__product_id__price_patch","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductSetPrice"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/products/{product_id}":{"delete":{"summary":"Delete Product","operationId":"delete_product_v1_products__product_id__delete","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/products/{product_id}/reserve":{"post":{"summary":"Reserve Product","operationId":"reserve_product_v1_products__product_id__reserve_post","parameters":[{"name":"product_id","in":"path","required":true,"schema":{"type":"integer","title":"Product Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductReservation"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/category/":{"post":{"summary":"Add Category","operationId":"add_category_v1_category__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/subcategory/":{"post":{"summary":"Add Sub Category","operationId":"add_sub_category_v1_subcategory__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubCategoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubCategoryBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/reservation/{reservation_id}/cancel_reserve":{"post":{"summary":"Cancel Reservation","operationId":"cancel_reservation_v1_reservation__reservation_id__cancel_reserve_post","parameters":[{"name":"reservation_id","in":"path","required":true,"schema":{"type":"integer","title":"Reservation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/reservation/{reservation_id}/confirm":{"post":{"summary":"Confirm Reservation","operationId":"confirm_reservation_v1_reservation__reservation_id__confirm_post","parameters":[{"name":"reservation_id","in":"path","required":true,"schema":{"type":"integer","title":"Reservation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/promotion/":{"post":{"summary":"Add Promotion","operationId":"add_promotion_v1_promotion__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromotionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PromotionBase"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CategoryBase":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"CategoryBase"},"CategoryCreate":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name","description":"The name of the category"}},"type":"object","required":["name"],"title":"CategoryCreate","example":{"name":"laptop"}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProductBase":{"properties":{"name":{"type":"string","title":"Name"},"price":{"type":"number","title":"Price"},"stock":{"type":"integer","title":"Stock"},"reserved_stock":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Reserved Stock","default":0},"subcategory_id":{"type":"integer","title":"Subcategory Id"}},"type":"object","required":["name","price","stock","subcategory_id"],"title":"ProductBase"},"ProductCreate":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name","description":"The name of the product"},"price":{"type":"integer","minimum":0.0,"title":"Price","description":"The price of the product in cents"},"stock":{"type":"integer","minimum":0.0,"title":"Stock","description":"The quantity available for sale"},"subcategory_id":{"type":"integer","minimum":0.0,"title":"Subcategory Id","description":"The id for subcategory"}},"type":"object","required":["name","price","stock","subcategory_id"],"title":"ProductCreate","example":{"name":"Lenovo T450","price":150,"stock":10,"subcategory_id":1}},"ProductReservation":{"properties":{"quantity":{"type":"integer","title":"Quantity"}},"type":"object","required":["quantity"],"title":"ProductReservation"},"ProductSetPrice":{"properties":{"price":{"type":"number","title":"Price"}},"type":"object","required":["price"],"title":"ProductSetPrice"},"PromotionBase":{"properties":{"status":{"type":"integer","title":"Status"},"discount_percentage":{"type":"number","title":"Discount Percentage"},"start_date":{"type":"string","format":"date-time","title":"Start Date"},"end_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Date"},"product_id":{"type":"integer","title":"Product Id"}},"type":"object","required":["status","discount_percentage","start_date","product_id"],"title":"PromotionBase"},"PromotionCreate":{"properties":{"discount_percentage":{"type":"number","maximum":100.0,"minimum":0.0,"title":"Discount Percentage"},"start_date":{"type":"string","format":"date-time","title":"Start Date","default":"2024-08-30T13:13:10.002419"},"end_date":{"type":"string","format":"date-time","title":"End Date"},"product_id":{"type":"integer","title":"Product Id"}},"type":"object","required":["discount_percentage","product_id"],"title":"PromotionCreate","example":{"discount_percentage":20.0,"end_date":"2024-10-01T12:00:00","product_id":1,"start_date":"2024-09-01T12:00:00"}},"ReservationBase":{"properties":{"id":{"type":"integer","title":"Id"},"product_id":{"type":"integer","title":"Product Id"},"status":{"type":"integer","title":"Status","default":0},"reserved_quantity":{"type":"integer","title":"Reserved Quantity"}},"type":"object","required":["id","product_id","reserved_quantity"],"title":"ReservationBase"},"SubCategoryBase":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"category_id":{"type":"integer","title":"Category Id"}},"type":"object","required":["id","name","category_id"],"title":"SubCategoryBase"},"SubCategoryCreate":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name","description":"The name of the sub category"},"category_id":{"type":"integer","title":"Category Id","description":"category id"}},"type":"object","required":["name","category_id"],"title":"SubCategoryCreate","example":{"category_id":1,"name":"Lenovo"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}